# Reduce input detection CPU usage