--- a/package.json
+++ b/package.json
@@ -18,17 +18,18 @@
         "sjcl-all": "^1.0.1"
     },
     "devDependencies": {
+        "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
         "@types/chrome": "^0.0.128",
         "@types/jquery": "^3.5.5",
         "@types/sjcl": "^1.0.29",
-        "css-loader": "^0.28.11",
-        "node-sass": "^5.0.0",
+        "css-loader": "^6.7.2",
+        "mini-css-extract-plugin": "^2.7.0",
+        "node-sass": "^8.0.0",
         "sass-loader": "^10.1.1",
         "style-loader": "^0.20.3",
         "svg-inline-loader": "^0.8.2",
         "ts-loader": "^8.0.14",
         "typescript": "^4.1.3",
-        "typings-for-css-modules-loader": "^1.7.0",
         "webpack": "^5.15.0",
         "webpack-cli": "^4.4.0"
     }