--- a/src/classes/PageControl.ts
+++ b/src/classes/PageControl.ts
@@ -29,77 +29,106 @@ export default class PageControl
     }
 
     /** Try to detect credentials fields */
-    public detectFields()
-    {
-        let passwordFields: JQuery = $('input[type="password"]');
-
-        if(passwordFields.length) // Found some password fields?
-        {
-            passwordFields.each((passwordIndex, passwordField)=>{ // Loop through password fields
-                this._createFieldSet(passwordField);
-            });
-        }
-
-        this._findCredentials();
-        this._attachEscapeEvent();
+    public detectFields() {
+        this.detectNewFields(document.querySelectorAll('input[type="password"]'));
     }
 
     /**
-     * Try to detect new credentials fields
+     * Try to detect new credentials fields.
      * @param passwordFields A list of changed or added password fields.
      */
-    public detectNewFields(passwordFields: NodeListOf<Element>)
-    {
-        for(const passwordField of passwordFields)
-        {
-            if(passwordField instanceof HTMLElement)
-                this._createFieldSet(passwordField);
+    public detectNewFields(passwordFields: NodeListOf<HTMLInputElement>) {
+        for (const passwordField of passwordFields) {
+            if (this._haveFieldSetForPasswordInput(passwordField)) {
+                continue;
+            }
+            const $passwordField = $(passwordField);
+            const isPasswordFieldVisible = $passwordField.is(':visible');
+            let input: HTMLInputElement | null = passwordField;
+            let foundFieldSet = false;
+            while ((input = PageControl._getPreviousInput(input)) != null) {
+                const inputType = input.type || 'text'; // Get input type, if none default to "text"
+                if ((inputType === 'text' || inputType === 'email' || inputType === 'tel') &&
+                    $passwordField.is(':visible') == isPasswordFieldVisible) {
+                    this._fieldSets.set(input, new FieldSet(this, $passwordField, $(input)));
+                    foundFieldSet = true;
+                    break;
+                }
+            }
+            if (!foundFieldSet && isPasswordFieldVisible) {
+                this._fieldSets.set(passwordField, new FieldSet(this, $passwordField, $(passwordField)));
+            }
         }
-
         this._findCredentials();
         this._attachEscapeEvent();
     }
 
     /**
-     * Create a fieldset for the `passwordField`. This method will also look for an username field
-     * @param passwordField The password field we're going to use
+     * Find the closest input element that appears before the given element in the DOM tree.
+     *
+     * @param element The starting element.
+     * @return The closest input element or null, if no input element was found before the starting element.
      */
-    private _createFieldSet(passwordField: HTMLElement)
-    {
-        let prevField: JQuery<HTMLElement> | undefined;
-        let prevVisibleField: JQuery<HTMLElement> | undefined;
-        let $passwordField = $(passwordField);
-        $('input').each((inputIndex, input) => { // Loop through input fields to find the field before our password field
-            const $input = $(input);
-            const inputType = $input.attr('type') || 'text'; // Get input type, if none default to "text"
-            if (inputType === 'text' || inputType === 'email' || inputType === 'tel') { // We didn't reach our password field?
-                prevField = $input; // Is this a possible username field?
-                if ($input.is(':visible')) {
-                    prevVisibleField = $input;
-                }
-            } else if (inputType === 'password' && $input.is($(passwordField))) { // Found our password field?
-                let controlField = $input.is(':visible') ? prevField : prevVisibleField; // When the passwordfield is visible, we don't care if the usernamefield is visible, otherwise we need a visible usernamefield
-                if (!controlField && $input.is(':visible')) {
-                    // We didn't find the username field. Check if password field is actually visible
-                    controlField = $passwordField;
-                } // Else we didn't find a visible username of password field
+    private static _getPreviousInput(element: Element): HTMLInputElement | null {
+        let currentElement : Element;
+        let elementsWithCheckedChildrenStack : Element[][] = [[]];
+        let childElement : Element | null;
+        let siblingElement : Element | null;
+        let parentElement : Element | null;
 
-                if(controlField)
-                {
-                    if(this._fieldSets.has(controlField[0])) // We already have a fieldset for this controlfield?
-                        this._fieldSets.get(controlField[0])!.passwordField = $passwordField; // Set the passwordfield again, it might have changed (This happens for Dropbox, see issue #86)
-                    else
-                        this._fieldSets.set(controlField[0], new FieldSet(this, $passwordField, controlField));
-                }
-
-                return false; // Break the each() loop
+        if ((siblingElement = element.previousElementSibling) !== null) {
+            currentElement = siblingElement;
+        } else {
+            if ((parentElement = element.parentElement) !== null) {
+                currentElement = parentElement;
+                elementsWithCheckedChildrenStack[0].push(parentElement);
+            } else {
+                return null;
             }
-        });
+        }
+        while (true) {
+            if (currentElement instanceof HTMLInputElement) {
+                return currentElement;
+            }
+            if (!elementsWithCheckedChildrenStack[elementsWithCheckedChildrenStack.length - 1]
+                .includes(currentElement)) {
+                if ((childElement = currentElement.lastElementChild) !== null) {
+                    elementsWithCheckedChildrenStack[elementsWithCheckedChildrenStack.length - 1].push(currentElement);
+                    elementsWithCheckedChildrenStack.push([]);
+                    currentElement = childElement;
+                    continue;
+                }
+            }
+            if ((siblingElement = currentElement.previousElementSibling) !== null) {
+                currentElement = siblingElement;
+            } else if ((parentElement = currentElement.parentElement) !== null) {
+                currentElement = parentElement;
+                elementsWithCheckedChildrenStack.pop();
+                if (elementsWithCheckedChildrenStack.length === 0) {
+                    elementsWithCheckedChildrenStack.push([parentElement]);
+                }
+            } else {
+                return null;
+            }
+        }
+    }
+
+    /**
+     * @param passwordField The password field to search for.
+     * @return Whether there is already a fieldset with the given password field.
+     */
+    private _haveFieldSetForPasswordInput(passwordField: HTMLInputElement): boolean {
+        for (const fieldSet of this._fieldSets.values()) {
+            if (fieldSet.passwordField.get(0) === passwordField) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private _attachEscapeEvent()
     {
-        if(this._installedEscapeHandler || !this._fieldSets || this._fieldSets.size === 0) {
+        if (this._installedEscapeHandler || this._fieldSets.size === 0) {
             return; // We're not going to listen to key presses if we don't need them
         }
         this._installedEscapeHandler = true;
@@ -112,8 +141,8 @@ export default class PageControl
 
     private _findCredentials()
     {
-        if(this._fieldSets.size) // We should only look for credentials if we found input fields for it
-        {
+        if (this._foundCredentials === undefined && this._fieldSets.size) {
+            // We should only look for credentials if we found input fields for it
             Client.findCredentials().then((credentials)=>{
                 this._foundCredentials = credentials;
                 this._fieldSets.forEach((fieldSet) => fieldSet.receivedCredentials());