--- a/src/content_script.ts
+++ b/src/content_script.ts
@@ -1,18 +1,18 @@
-import * as $ from 'jquery-slim';
+import $ from 'jquery-slim';
 
 import PageControl from './classes/PageControl';
 const pageControl = new PageControl();
 
 $(()=>{
     pageControl.detectFields();
-    
+
     if(pageControl.settings.searchForInputsOnUpdate) // Should the MutationObserver be enabled?
     {
         // The MutationObserver implementation
         const observer = new MutationObserver((mutations)=>{
             for(const mutation of mutations)
             {
-                for(const node of mutation.addedNodes) 
+                for(const node of mutation.addedNodes)
                 {
                     if(node instanceof HTMLElement)
                     {