--- a/src/credentialSelector.ts
+++ b/src/credentialSelector.ts
@@ -1,4 +1,4 @@
-import * as $ from 'jquery-slim';
+import $ from 'jquery-slim';
 import * as IMessage from './IMessage';
 
 $(()=>{
@@ -14,9 +14,9 @@ class CredentialSelector
     {
         const params = new URLSearchParams(location.search);
 
-        this._nonce = parseInt(params.get('nonce')!);    
+        this._nonce = parseInt(params.get('nonce')!);
         this._credentialsList = $('#credentialsList');
-        
+
         this._fetchCredentials();
     }