--- a/src/popup.ts
+++ b/src/popup.ts
@@ -1,4 +1,4 @@
-import * as $ from 'jquery-slim';
+import $ from 'jquery-slim';
 import Client from './classes/BackgroundClient';
 
 $(()=>{