… dist/manifest.json