@Abestanis 2022-11-29
@RoelVB This is still valid and it still reduces the runtime of the `detectFields` function.
It includes the changes form #90, so this should be merged after it.