@Abestanis 2022-11-29
The timings were gathered with the following code changes to `content_script.ts`:
```ts
import $ from 'jquery-slim';

import PageControl from './classes/PageControl';
const pageControl = new PageControl();

$(()=>{
    console.time('detectFields')
    pageControl.detectFields();
    console.timeEnd('detectFields');
    if (pageControl.settings.searchForInputsOnUpdate) // Should the MutationObserver be enabled?
    {
        // The MutationObserver implementation
        const observer = new MutationObserver((mutations) => {
            if (mutations.length == 11) {
                console.time('ChangeObserver')
            }
            for(const mutation of mutations)
            {
                for(const node of mutation.addedNodes)
                {
                    if(node instanceof HTMLElement)
                    {
                        const passwordFields = node.querySelectorAll('input[type="password"]');

                        if(passwordFields.length)
                            setTimeout(() => pageControl.detectNewFields(passwordFields), 100);
                    }
                }
            }
            if (mutations.length == 11) {
                console.timeEnd('ChangeObserver');
            }
        });

        // Start observing
        observer.observe(document.body, {
            childList: true,
            subtree: true,
        });
    }
});
```
The test was performed on https://www.virustotal.com/gui/sign-in by reloading the page multiple times. The unit is milliseconds.

| detectFields [Before] | ChangeObserver [Before] | detectFields [After] | ChangeObserver [After] |
|------------------------|-----------------------------|----------------------|----------------------------|
0.40478515625 | 0.098876953125 | 0.380859375 | 0.137939453125
0.447021484375 | 0.087890625 | 0.13916015625 | 0.059814453125
0.571044921875 | 0.09716796875 | 0.181884765625 | 0.048828125
0.43115234375 | 0.10498046875 | 0.133056640625 | 0.107177734375
0.491943359375 | 0.06103515625 | 0.18505859375 | 0.095703125
0.419921875 | 0.080078125 | 0.202880859375 | 0.15869140625
0.35009765625 | 0.090087890625 | 0.14404296875 | 0.048828125
0.386962890625 | 0.31884765625 | 0.155029296875 | 0.06103515625
0.257080078125 | 0.12109375 | 0.141845703125 | 0.094970703125
0.358154296875 | 0.133056640625 | 0.22802734375 | 0.06298828125
0.365966796875 | 0.0869140625 | 0.13916015625 | 0.114013671875
0.366943359375 | 0.17724609375 | 0.134033203125 | 0.10205078125
0.3779296875 | 0.08984375 | 0.145751953125 | 0.04296875
0.2880859375 | 0.176025390625 | 0.14111328125 | 0.0458984375
0.464111328125 | 0.14892578125 | 0.138916015625 | 0.109130859375
0.359130859375 | 0.063720703125 | 0.202880859375 | 0.06689453125
0.262939453125 | 0.097900390625 | 0.125244140625 | 0.052978515625
0.284912109375 | 0.26611328125 | 0.14306640625 | 0.046875
0.3701171875 | 0.114013671875 | |
0.546875 | 0.051025390625 | |
0.275146484375 | 0.120849609375 | |

The average can be seen in the following table:

| detectFields [Before] | ChangeObserver [Before] | detectFields [After] | ChangeObserver [After] |
|------------------------|-----------------------------|----------------------|----------------------------|
| 0.3848 | 0.1231 | 0.1701 | 0.0809 |