@RoelVB 2022-12-11
I've made a version of CKP that fully got rid of jQuery a long time ago, which improves performance.
But I decided to not release it, because it contains changes that I don't fully trust to work for 100% and I was planning on rewriting a major part of CKP anyhow.

Also an important part of the rewrite is adding automated tests. Which would make me less hesitant to publish changes like completely getting rid of jQuery. I think we might also get a bèta release of CKP, so we can publish big changes into the bèta first.

If you don't mind I would like to close this PR, since similar/overlapping changes are included in the rewrite.