@Abestanis 2022-12-11
> I've made a version of CKP that fully got rid of jQuery

Nice, I hope my other pull requests don't conflict with too much, but luckily they are mostly on the backend, not using JQuery.

> Also an important part of the rewrite is adding automated tests.

That is awesome! 💯

> If you don't mind I would like to close this PR, since similar/overlapping changes are included in the rewrite.

No, makes sense, although it might be good to see if the general idea of this pr can still be applied to your branch in the future.
The main improvement comes from not searching through all inputs on the site for each password field, instead just look for the closest input element before the password input by walking the html element tree. The main function, `_getPreviousInput`, doesn't use any JQuery if I see that correctly, so you could just copy it over to your branch.